requirements.txt
# LearningAPI
django>=2.2.1
djangorestframework>=3.9.3
gunicorn>=19.9.0
pylint>=2.3.1
pylint-django>=2.0.9
